chore(sample): move sample app to Samples folder
