--- JsonStores.Sample/JsonStores.Sample.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\JsonStores\JsonStores.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <PackageReference Include="System.CommandLine" Version="2.0.0-beta1.20574.7" />
-        <PackageReference Include="System.CommandLine.Rendering" Version="0.3.0-alpha.20574.7" />
-    </ItemGroup>
-</Project>
