+++ Samples/SampleApp/ListView.cs
@@ -4,7 +4,7 @@ using System.CommandLine.Rendering.Views;
 using System.Linq;
 using System.Reflection;
 
-namespace JsonStores.Sample
+namespace JsonStores.Samples.SampleApp
 {
     public sealed class ListView<T> : StackLayoutView where T : class, new()
     {
