+++ Samples/SampleApp/Models/Country.cs
@@ -1,0 +1,12 @@
+﻿using JsonStores.Annotations;
+
+namespace JsonStores.Samples.SampleApp.Models
+{
+    public class Country
+    {
+        [JsonRepositoryKey]
+        public int CountryCode { get; set; }
+
+        public string Name { get; set; }
+    }
+}
