+++ Samples/SampleApp/README.md
@@ -1,0 +1,7 @@
+# JSON Stores Sample App
+
+Basic usage using [System.CommandLine](https://github.com/dotnet/command-line-api) API
+
+## Instructions
+
+Use `sample --help` to get the availabe commands
